pyproject.toml
    [project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.26.0",
    "numpy>=2.2.3",
    "openai-clip>=1.0.1",
    "pillow>=11.1.0",
    "protobuf>=5.29.3",
    "sentencepiece>=0.2.0",
    "sqlite-vec>=0.1.6",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
]
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

    [tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
transformers = { git = "https://github.com/huggingface/transformers.git" }
